+++ lib/localize/utils.ex
@@ -43,14 +43,11 @@ defmodule Localize.Utils do
     major = :erlang.system_info(:otp_release) |> List.to_string()
     vsn_file = Path.join([:code.root_dir(), "releases", major, "OTP_VERSION"])
 
-    try do
-      {:ok, contents} = File.read(vsn_file)
-      String.split(contents, "\n", trim: true)
+    with {:ok, contents} <- File.read(vsn_file),
+         [full] <- String.split(contents, "\n", trim: true) do
+      full
     else
-      [full] -> full
-      _ -> major
-    catch
-      :error, _ -> major
+      _no_full_version -> major
     end
   end
 end